Moved Settings dialog components to folders and files
