+++ shortcut_composer/composer_utils/layouts/buttons_layout.py
