+++ shortcut_composer/composer_utils/tabs/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: © 2022 Wojciech Trybus <wojtryb@gmail.com>
+# SPDX-License-Identifier: GPL-3.0-or-later
+
+from .general_settings_tab import GeneralSettingsTab
+from .pie_values_tab import PieValuesTab
+
+
+__all__ = ["GeneralSettingsTab", "PieValuesTab"]
